main rejoin: OpInfo tests (#398): python banked, F64 becomes a real executable dtype

+++ crates/luminal_python/rust/src/translator/attention.rs
@@ -76,7 +76,7 @@ impl<'a> Translator<'a> {
         // Default scale 1/sqrt(head_dim) needs a concrete head_dim;
         // symbolic-shape HF graphs always pass `scale` explicitly.
         let scale = match float_arg("scale") {
-            Some(v) => v as f32,
+            Some(v) => v,
             None => {
                 let head_dim = query
                     .legacy_tracker_ref()
@@ -87,7 +87,7 @@ impl<'a> Translator<'a> {
                         "SDPA: query head_dim must be concrete to derive the default \
                          scale (pass `scale` explicitly for symbolic head dims)",
                     )?;
-                1.0_f32 / (head_dim as f32).sqrt()
+                1.0_f64 / (head_dim as f64).sqrt()
             }
         };
